move tests location
