+++ core/node/da_clients/src/eigen/mod.rs
@@ -1,9 +1,7 @@
 mod blob_info;
 mod client;
-mod client_tests;
 mod sdk;
 mod verifier;
-mod verifier_tests;
 
 pub use self::client::{EigenClient, GetBlobData};
 #[allow(clippy::all)]
